+++ Assets/Scripts/ServerConnection/Aiortc/AiortcConnector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e1f192c59c54fbd87c25983cfdf6704
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
