--- Assets/Scripts/ServerConnection/Aiortc/aiortcConnector.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 0e1f192c59c54fbd87c25983cfdf6704
-timeCreated: 1672842185
